chapter2/src/test/java/org/smart4j/chapter2/service/CustomerServiceTest.java
package org.smart4j.chapter2.service;

import lombok.NoArgsConstructor;
import org.junit.Assert;
import org.junit.Before;
import org.junit.Test;
import org.smart4j.chapter2.model.Customer;

import java.util.HashMap;
import java.util.List;
import java.util.Map;

import static org.junit.Assert.*;

/**
 * CustomerService单元测试
 */
public class CustomerServiceTest {
    private final CustomerService customerService = new CustomerService();

//    public CustomerServiceTest(CustomerService customerService) {
//        this.customerService = customerService;
//    }

    @Before
    public void init() {
        // TODO: 2022/12/12 初始化数据库
    }

    @Test
    public void getCustomerList() {
        List<Customer> customerList = customerService.getCustomerList();
        assertEquals(2, customerList.size());
    }

    @Test
    public void getCustomer() {
        long id = 1;
        Customer customer = customerService.getCustomer(id);
        assertNotNull(customer);
    }

    @Test
    public void createCustomer() {
        Map<String, Object> fieldMap = new HashMap<String, Object>();
        fieldMap.put("name", "customer100");
        fieldMap.put("contact", "John");
        fieldMap.put("telephone", "1321541241");
        boolean result = customerService.createCustomer(fieldMap);

        assertTrue(result);
    }

    @Test
    public void updateCustomer() {
        long id = 1;
        Map<String, Object> fieldMap = new HashMap<String, Object>();
        fieldMap.put("contact", "Eric");
        boolean result = customerService.updateCustomer(id, fieldMap);

        assertTrue(result);
    }

    @Test
    public void deleteCustomer() {
        long id = 1;
        boolean result = customerService.deleteCustomer(id);

        assertTrue(result);
    }
}